Fix and extension
For me the last valid search never returned any values, so I changed the
regExp. My other issue was that jobs had to be added one by one, and I
did not want to make tens of Nagios checks for every single job, just
one that automatically updates, if new job is added, so I extended the
script with a query to  get all jobs that are scheduled for the next 10
days, and check their last status and only show those results that has
problems.